style(users): adjust styling w/ huge enhancements
