--- src/components/EmployeeTable/EmployeeTableRow.module.css
@@ -1,117 +1,0 @@
-.employee__management__name__cell {
-  display: flex;
-}
-
-.employee__management__avatar {
-  margin-right: 10px;
-  width: 30px;
-  height: 30px;
-  border-radius: var(--border-round);
-  background-color: var(--primary-white-2);
-  overflow: hidden;
-}
-
-.employee__management__name__email__container {
-  display: flex;
-  flex-direction: column;
-}
-
-.employee__management__email {
-  font-size: var(--font-size);
-  color: var(--primary-black-4);
-}
-
-.employee__management__role__superadmin {
-  background-color: var(--badge-danger-color);
-  width: fit-content;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  padding: 3px 5px;
-  color: var(--primary-white);
-  border-radius: var(--border-radius);
-  font-size: var(--font-size);
-}
-
-.employee__management__role__employee {
-  background-color: var(--badge-primary-color);
-  width: fit-content;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  color: var(--primary-white);
-  padding: 3px 5px;
-  border-radius: var(--border-radius);
-  font-size: var(--font-size);
-}
-
-.employee__management__role__hradmin {
-  background-color: var(--badge-warning-color);
-  width: fit-content;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  color: var(--primary-white);
-  padding: 3px 5px;
-  border-radius: var(--border-radius);
-  font-size: 14px;
-}
-
-.employee__management__role__admin {
-  background-color: var(--badge-info-color);
-  width: fit-content;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  color: var(--primary-white);
-  padding: 3px 5px;
-  border-radius: var(--border-radius);
-  font-size: 14px;
-}
-
-.employee__management__avatar img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }  
-
-.employee__management__action__container {
-  display: flex;
-  flex-direction: row;
-}
-
-.employee__management__action {
-  margin-right: 5px;
-  padding: 8px 10px;
-  color: var(--primary-white);
-  background-color: #2980b9;
-  border: none;
-  border-radius: var(--border-radius);
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.employee__management__action__view {
-  background-color: #8e44ad;
-  text-decoration: none;
-}
-
-.employee__management__action__view:hover {
-  background-color: #6c3483;
-}
-
-.employee__management__action__edit {
-  background-color: #f39c12;
-}
-
-.employee__management__action__edit:hover {
-  background-color: #d68910;
-}
-
-.employee__management__action__delete {
-  background-color: #e74c3c;
-}
-
-.employee-management__action__delete:hover {
-  background-color: #c0392b;
-}
