+++ src/feature/Users/components/EmployeeTable/EmployeeTable.module.css
@@ -1,0 +1,21 @@
+.table {
+  width: 100%;
+  border-collapse: collapse;
+  margin-top: 20px;
+  width: 100%;
+  text-align: left;
+}
+
+.table th {
+  font-weight: var(--font-bold-400);
+}
+
+.table th,
+.table td {
+  padding: 10px;
+  text-align: left;
+}
+
+.table td {
+  border-top: 1px solid #dee2e6;
+}
